[Test] Make conftests only depend on local data if possible.
Data for conftests that testing tsv files is stored locally. Reference data is added to the repo. Binary files are still stored on MPIB datashare but the tests are skipped gracefully. This is ok as the integration tests for binary data are typically backed up by tests with tsv data